\name{affmat}
\alias{affmat}
\title{Objects count by site to affiliation matrix}
\usage{affmat()}
\description{This function takes an input of object counts by archaeological site and produces two affiliation matrices as output. The first matrix is a weighted affiliation matrix that indicates the strength of the relationship between each site and each object type based on the count of objects found at each site. The second matrix is a binary affiliation matrix that indicates only the presence (1) or absence (0) of each object type at each site. These matrices can be useful for exploring patterns of object distribution and site similarity in archaeological data.}
